admin method done

+++ src/pages/Users/User/User.js
@@ -9,19 +9,22 @@ const User = (props) => {
 
 
     const handleDelete = id => {
-        const url = `https://aqueous-badlands-20033.herokuapp.com/users/${id}`;
-        fetch(url, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount) {
-                    const remaining = users.filter(user => user._id !== id);
-                    setUsers(remaining);
+        const result = window.confirm('Want to remove this item?');
 
-                    alert('deleted successfully');
-                }
-            });
+        if (result) {
+            const url = `https://aqueous-badlands-20033.herokuapp.com/users/${id}`;
+            fetch(url, {
+                method: 'DELETE'
+            })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount) {
+                        const remaining = users.filter(user => user._id !== id);
+                        setUsers(remaining);
+                        alert('deleted successfully');
+                    }
+                });
+        }
     }
 
     return (
